Refactor: Move notification channel
