+++ src/components/notification/channel/webhook.ts
@@ -1,6 +1,6 @@
 import axios from 'axios'
 
-import { WebhookData } from '../../interfaces/data'
+import { WebhookData } from '../../../interfaces/data'
 
 export const sendWebhook = async (data: WebhookData) => {
   try {
